decoupled PreparedBoundaries
